+++ CESCSAP-back/CESCSAP/src/main/java/com/mhy/cescsap/service/LogService.java
@@ -11,13 +11,13 @@ import java.util.List;
 public interface LogService {
     //记录操作日志
     //查询一条日志
-    Log getLogById(Integer logId);
+    Log getLogById(Long logId);
     //查询所有日志
     List<Log> getAllLogs();
     //添加一条日志
     Integer addLog(Log log);
     //删除一条日志
-    Integer deleteLog(Integer logId);
+    Integer deleteLog(Long logId);
     //修改一条日志
     Integer updateLog(Log log);
 }
